+++ src/anomalous-value/anomalous-value.spec.js
@@ -10,7 +10,7 @@ const {
 
 // local definitions
 const { LendingPool: address } = require('../../contract-addresses.json');
-const { abi } = require('../../interfaces/ILendingPool.json');
+const { abi } = require('../../abi/ILendingPool.json');
 const { handleTransaction } = require('./anomalous-value');
 
 // create interface
